+++ tunic/install.py
@@ -52,7 +52,8 @@ class VirtualEnvInstallation(ProjectBaseMixin):
         alternative sources from which to download dependencies and path to the
         virtualenv tool.
 
-        :param str base: Absolute path to the root of the code deploy
+        :param str base: Absolute path to the root of the code deploy on
+            the remote server
         :param list packages: A collection of package names to install into
             a remote virtual environment.
         :param list sources: A collection of alternative sources from which
@@ -155,10 +156,12 @@ class StaticFileInstallation(ProjectBaseMixin):
     """
 
     def __init__(self, base, local_path, runner=None):
-        """Set the project base directory and path to a directory of static
-        content to be installed into a remote release directory.
+        """Set the project base directory on the remote server and path to a
+        directory of static content to be installed into a remote release
+        directory.
 
-        :param str base: Absolute path to the root of the code deploy
+        :param str base: Absolute path to the root of the code deploy on
+            the remote server
         :param str local_path: Absolute or relative path to a local directory
             whose contents will be copied to a remote release directory.
         :param FabRunner runner: Optional runner to use for executing
@@ -190,7 +193,7 @@ class StaticFileInstallation(ProjectBaseMixin):
             deployment. If this ID corresponds to a directory that already
             exists, contents of the local directory will be copied into
             this directory.
-        :return: The results of the upload command using Fabric. This return
+        :return: The results of the ``put`` command using Fabric. This return
             value is an iterable of the paths of all files uploaded on the remote
             server.
         """
@@ -203,6 +206,80 @@ class StaticFileInstallation(ProjectBaseMixin):
         # put command.
         local_path = self._local_path.strip('*').strip(os.path.sep)
         return self._runner.put(os.path.join(local_path, '*'), release_path)
+
+
+class LocalArtifactInstallation(ProjectBaseMixin):
+    """Install a single local file into a remote release directory.
+
+    This can be useful for installing applications that are typically bundled
+    as a single file, e.g. Go binaries, Java JAR or WAR files, etc.. The artifact
+    can optionally be renamed as part of the installation process.
+
+    If the remote release directory does not already exist, it will be
+    created during the install process.
+
+    See :doc:`design` for more information about the expected directory
+    structure for deployments.
+
+    .. versionadded:: 1.1.0
+    """
+
+    def __init__(self, base, local_file, remote_name=None, runner=None):
+        """Set the project base directory on the remote server, local artifact (a
+        single file) that should be installed remotely, and optional file name
+        to rename the artifact to on the remote server.
+
+        :param str base: Absolute path to the root of the code deploy on
+            the remote server
+        :param str local_file: Relative or absolute path to the local artifact
+            to be installed on the remote server.
+        :param str remote_name: Optional file name for the artifact after it has
+            been installed on the remote server. For example, if the artifact should
+            always be called 'application.jar' on the remote server but might
+            be named differently ('application-1.2.3.jar') locally, you would
+            specify ``remote_name='application.jar'`` for this parameter.
+        :param FabRunner runner: Optional runner to use for executing
+            remote and local commands to perform the installation.
+        :raises ValueError: If the base directory or local file isn't
+            specified.
+        """
+        super(LocalArtifactInstallation, self).__init__(base)
+
+        if not local_file:
+            raise ValueError("You must specify a local file path")
+
+        self._local_file = local_file
+        self._remote_name = remote_name
+        self._runner = runner if runner is not None else FabRunner()
+
+    def install(self, release_id):
+        """Install the local artifact into the remote release directory, optionally
+        with a different name than the artifact had locally.
+
+        If the directory for the given release ID does not exist on the remote
+        system, it will be created. The directory will be created according to
+        the standard Tunic directory structure (see :doc:`design`).
+
+        :param str release_id: Timestamp-based identifier for this
+            deployment.
+        :return: The results of the ``put`` command using Fabric. This return
+            value is an iterable of the paths of all files uploaded on the remote
+            server.
+        """
+        release_path = os.path.join(self._releases, release_id)
+        if not self._runner.exists(release_path):
+            self._runner.run("mkdir -p '{0}'".format(release_path))
+
+        # The artifact can optionally be renamed when being uploaded to
+        # remote server. Useful for when we need a consistent name for
+        # each deploy on the remote server but the local artifact includes
+        # version numbers or something.
+        if self._remote_name is not None:
+            destination = os.path.join(release_path, self._remote_name)
+        else:
+            destination = release_path
+
+        return self._runner.put(self._local_file, destination)
 
 
 # pylint: disable=too-few-public-methods